Move Cert Authority methods out of api to remove dependency on crypto/ssh.
