+++ integration/integration_test.go
@@ -3898,11 +3898,11 @@ func (s *IntSuite) TestRotateChangeSigningAlg(c *check.C) {
 	assertSigningAlg := func(svc *service.TeleportProcess, alg string) {
 		hostCA, err := svc.GetAuthServer().GetCertAuthority(services.CertAuthID{Type: services.HostCA, DomainName: Site}, false)
 		c.Assert(err, check.IsNil)
-		c.Assert(hostCA.GetSigningAlg(), check.Equals, alg)
+		c.Assert(services.GetSigningAlgName(hostCA), check.Equals, alg)
 
 		userCA, err := svc.GetAuthServer().GetCertAuthority(services.CertAuthID{Type: services.UserCA, DomainName: Site}, false)
 		c.Assert(err, check.IsNil)
-		c.Assert(userCA.GetSigningAlg(), check.Equals, alg)
+		c.Assert(services.GetSigningAlgName(userCA), check.Equals, alg)
 	}
 
 	rotate := func(svc *service.TeleportProcess, mode string) *service.TeleportProcess {
